requirements.txt
# ============================================================
# Infra-Pilot — Python Dependencies
# Install with: pip install -r requirements.txt
# ============================================================

# ── Web Framework (Backend API) ────────────────────────────
fastapi>=0.110.0,<0.115.0
uvicorn[standard]>=0.28.0,<0.32.0
python-multipart>=0.0.9

# ── Environment & Config ───────────────────────────────────
python-dotenv>=1.0.0

# ── Git / GitHub Integration ───────────────────────────────
gitpython>=3.1.40
PyGithub>=2.3.0

# ── File Parsing ───────────────────────────────────────────
PyYAML>=6.0.1
python-hcl2>=4.3.0

# ── Drift Detection ────────────────────────────────────────
deepdiff>=7.0.1

# ── AI — Local (Ollama) ────────────────────────────────────
ollama>=0.2.0

# ── AI — Cloud (Google Gemini) ─────────────────────────────
google-generativeai>=0.7.0

# ── Delivery — Google Sheets ───────────────────────────────
gspread>=6.1.0
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0

# ── Delivery — Email ───────────────────────────────────────
yagmail>=0.15.0

# ── Utilities ──────────────────────────────────────────────
httpx>=0.27.0
pydantic>=2.9.0
tenacity>=8.3.0
